+++ apps/extension/src/core/domains/app/handler.ts
@@ -1,4 +1,4 @@
-import { DEBUG } from "@core/constants"
+import { DEBUG, TALISMAN_WEB_APP_DOMAIN } from "@core/constants"
 import { AccountMeta } from "@core/domains/accounts/types"
 import { AppStoreData } from "@core/domains/app/store.app"
 import type {
@@ -41,6 +41,17 @@ export default class AppHandler extends ExtensionHandler {
 
     const account = this.getRootAccount()
     assert(!account, "A root account already exists")
+
+    // Before any accounts are created, we want to add talisman.xyz as an authorised site with connectAllSubstrate
+    this.stores.sites.set({
+      [TALISMAN_WEB_APP_DOMAIN]: {
+        addresses: [],
+        connectAllSubstrate: true,
+        id: TALISMAN_WEB_APP_DOMAIN,
+        origin: "Talisman",
+        url: `https://${TALISMAN_WEB_APP_DOMAIN}`,
+      },
+    })
 
     let confirmed = false
     const method = mnemonic ? "import" : "new"
@@ -189,6 +200,19 @@ export default class AppHandler extends ExtensionHandler {
     return true
   }
 
+  private async resetWallet() {
+    // delete all the accounts
+    keyring.getAccounts().forEach((acc) => keyring.forgetAccount(acc.address))
+    this.stores.app.set({ onboarded: "FALSE" })
+    await this.stores.password.reset()
+    await this.stores.seedPhrase.clear()
+    await this.state.openOnboarding("/import?resetWallet=true")
+    // since all accounts are being wiped, all sites need to be reset - so they may as well be wiped.
+    await this.stores.sites.clear()
+
+    return true
+  }
+
   private async dashboardOpen({ route }: RequestRoute): Promise<boolean> {
     if (!(await this.stores.app.getIsOnboarded())) return this.onboardOpen()
     this.state.openDashboard({ route })
@@ -296,6 +320,9 @@ export default class AppHandler extends ExtensionHandler {
           (request as RequestTypes["pri(app.phishing.addException)"]).url
         )
       }
+
+      case "pri(app.resetWallet)":
+        return this.resetWallet()
 
       default:
         throw new Error(`Unable to handle message of type ${type}`)