+++ apps/extension/src/core/domains/sitesAuthorised/types.ts
@@ -54,14 +54,18 @@ export type AuthorizedSite = {
   origin: string
   url: string
   ethChainId?: number
+  connectAllSubstrate?: boolean
 }
 
 export type ProviderType = "polkadot" | "ethereum"
 
+export declare type AuthorisedSiteUpdate = Omit<Partial<AuthorizedSite>, "id">
+
 export declare type RequestAuthorizedSiteUpdate = {
   id: string
-  props: Omit<Partial<AuthorizedSite>, "id">
+  authorisedSite: AuthorisedSiteUpdate
 }
+
 export declare type RequestAuthorizedSiteForget = { id: string; type: ProviderType }
 
 export interface AuthorizeRequest {