+++ apps/extension/src/core/handlers/State.ts
@@ -11,7 +11,6 @@ import { SigningRequestsStore } from "@core/domains/signing"
 import { SitesRequestsStore, sitesAuthorisationStore } from "@core/domains/sitesAuthorised"
 import EvmWatchAssetRequestsStore from "@core/domains/tokens/evmWatchAssetRequestsStore"
 import { isEthereumRequest } from "@core/util/isEthereumRequest"
-import { assert } from "@polkadot/util"
 import { sleep } from "@talismn/util"
 import Browser from "webextension-polyfill"
 
@@ -214,14 +213,18 @@ export default class State {
     return tab
   }
 
-  public async openOnboarding() {
+  public async openOnboarding(route?: string) {
     if (this.#onboardingTabOpening) return
     this.#onboardingTabOpening = true
-    const url = Browser.runtime.getURL(`onboarding.html`)
+    const baseUrl = Browser.runtime.getURL(`onboarding.html`)
 
     const onboarded = await appStore.getIsOnboarded()
 
-    await this.openTabOnce({ url, shouldFocus: onboarded })
+    await this.openTabOnce({
+      url: `${baseUrl}${route ? `#${route}` : ""}`,
+      baseUrl,
+      shouldFocus: onboarded,
+    })
     this.#onboardingTabOpening = false
   }