+++ apps/extension/src/ui/api/types.ts
@@ -29,6 +29,7 @@ import { AnySigningRequest, TransactionDetails } from "@core/domains/signing/typ
 import {
   AuthRequestAddresses,
   AuthRequestId,
+  AuthorisedSiteUpdate,
   AuthorizeRequest,
   AuthorizedSite,
   AuthorizedSites,
@@ -88,6 +89,7 @@ export default interface MessageTypes {
   modalOpen: (modal: ModalOpenRequest) => Promise<boolean>
   modalOpenSubscribe: (cb: (val: ModalOpenRequest) => void) => UnsubscribeFn
   analyticsCapture: (request: AnalyticsCaptureRequest) => Promise<boolean>
+  resetWallet: () => Promise<boolean>
 
   // mnemonic message types -------------------------------------------------------
   mnemonicUnlock: (pass: string) => Promise<string>
@@ -115,7 +117,12 @@ export default interface MessageTypes {
   accountValidateMnemonic: (mnemonic: string) => Promise<boolean>
 
   // balance message types ---------------------------------------------------
-  getBalance: ({ chainId, evmNetworkId, tokenId, address }: RequestBalance) => Promise<BalanceJson>
+  getBalance: ({
+    chainId,
+    evmNetworkId,
+    tokenId,
+    address,
+  }: RequestBalance) => Promise<BalanceJson | undefined>
   getBalanceLocks: ({ chainId, addresses }: RequestBalanceLocks) => Promise<ResponseBalanceLocks>
   balances: (cb: () => void) => UnsubscribeFn
   balancesByParams: (
@@ -130,10 +137,7 @@ export default interface MessageTypes {
   authorizedSite: (id: string) => Promise<AuthorizedSite>
   authorizedSiteSubscribe: (id: string, cb: (sites: AuthorizedSite) => void) => UnsubscribeFn
   authorizedSiteForget: (id: string, type: ProviderType) => Promise<boolean>
-  authorizedSiteUpdate: (
-    id: string,
-    properties: Omit<Partial<AuthorizedSite>, "id">
-  ) => Promise<boolean>
+  authorizedSiteUpdate: (id: string, authorisedSite: AuthorisedSiteUpdate) => Promise<boolean>
 
   // authorization requests message types ------------------------------------
   authRequestsSubscribe: (cb: (requests: AuthorizeRequest[]) => void) => UnsubscribeFn