+++ apps/extension/src/ui/apps/popup/pages/Sign/EthSignTransactionRequest.tsx
@@ -1,18 +1,17 @@
 import { AccountJsonHardwareEthereum } from "@core/domains/accounts/types"
-import { BalanceFormatter } from "@core/domains/balances"
+import { EthPriorityOptionName } from "@core/domains/signing/types"
 import { AppPill } from "@talisman/components/AppPill"
 import Grid from "@talisman/components/Grid"
 import { SimpleButton } from "@talisman/components/SimpleButton"
+import { WithTooltip } from "@talisman/components/Tooltip"
+import { InfoIcon } from "@talisman/theme/icons"
 import { Content, Footer, Header } from "@ui/apps/popup/Layout"
-import Fiat from "@ui/domains/Asset/Fiat"
-import Tokens from "@ui/domains/Asset/Tokens"
-import { EthFeeSelect } from "@ui/domains/Ethereum/EthFeeSelect"
+import { TokensAndFiat } from "@ui/domains/Asset/TokensAndFiat"
+import { EthFeeSelect } from "@ui/domains/Ethereum/GasSettings/EthFeeSelect"
 import { EthSignBody } from "@ui/domains/Ethereum/Sign/EthSignBody"
 import { SignAlertMessage } from "@ui/domains/Ethereum/Sign/shared"
 import { useEthSignTransactionRequest } from "@ui/domains/Sign/SignRequestContext"
-import useToken from "@ui/hooks/useToken"
-import { useTokenRates } from "@ui/hooks/useTokenRates"
-import { BigNumber } from "ethers"
+import { useBalance } from "@ui/hooks/useBalance"
 import { Suspense, lazy, useCallback, useEffect, useMemo } from "react"
 import styled from "styled-components"
 import { Button } from "talisman-ui"
@@ -96,6 +95,56 @@ const SignContainer = styled(Container)`
   }
 `
 
+const FeeTooltip = ({
+  account,
+  estimatedFee,
+  maxFee,
+  tokenId,
+}: {
+  account?: string
+  estimatedFee?: string | bigint
+  maxFee?: string | bigint
+  tokenId?: string
+}) => {
+  const balance = useBalance(account as string, tokenId as string)
+
+  if (!estimatedFee && !maxFee && !balance) return null
+  return (
+    <div className="flex flex-col gap-2 whitespace-nowrap text-sm">
+      {!!estimatedFee && (
+        <div className="flex w-full justify-between gap-8">
+          <div>Estimated Fee:</div>
+          <div>
+            <TokensAndFiat tokenId={tokenId} planck={estimatedFee} noTooltip noCountUp />
+          </div>
+        </div>
+      )}
+      {!!maxFee && (
+        <div className="flex w-full justify-between gap-8">
+          <div>Max Fee:</div>
+          <div>
+            <TokensAndFiat tokenId={tokenId} planck={maxFee} noTooltip noCountUp />
+          </div>
+        </div>
+      )}
+      {!!balance && (
+        <div className="flex w-full justify-between gap-8">
+          <div>Balance:</div>
+          <div>
+            <TokensAndFiat
+              tokenId={tokenId}
+              planck={balance.free.planck}
+              noTooltip
+              noCountUp
+              isBalance
+            />
+          </div>
+        </div>
+      )}
+    </div>
+  )
+}
+
 export const EthSignTransactionRequest = () => {
   const {
     url,
@@ -116,6 +165,11 @@ export const EthSignTransactionRequest = () => {
     isPayloadLocked,
     setIsPayloadLocked,
     transactionInfo,
+    gasSettingsByPriority,
+    setCustomSettings,
+    setReady,
+    isValid,
+    networkUsage,
   } = useEthSignTransactionRequest()
 
   const { processing, errorMessage } = useMemo(() => {
@@ -130,29 +184,22 @@ export const EthSignTransactionRequest = () => {
     if (status === "SUCCESS") window.close()
   }, [status])
 
-  const nativeToken = useToken(network?.nativeToken?.id)
-  const nativeTokenRates = useTokenRates(nativeToken?.id)
-
   // gas settings must be locked as soon as payload is sent to ledger
   const handleSendToLedger = useCallback(() => {
     setIsPayloadLocked(true)
   }, [setIsPayloadLocked])
 
-  const estimatedFee = useMemo(
-    () =>
-      txDetails && nativeToken
-        ? new BalanceFormatter(
-            BigNumber.from(txDetails?.estimatedFee).toString(),
-            nativeToken?.decimals,
-            nativeTokenRates
-          )
-        : null,
-    [nativeToken, nativeTokenRates, txDetails]
+  const isReadyToDisplay = useMemo(
+    () => Boolean(transactionInfo && (txDetails?.estimatedFee || errorMessage)),
+    [transactionInfo, txDetails?.estimatedFee, errorMessage]
   )
 
-  const isReadyToDisplay = useMemo(
-    () => Boolean(transactionInfo && (estimatedFee || errorMessage)),
-    [transactionInfo, estimatedFee, errorMessage]
+  const handleFeeChange = useCallback(
+    (priority: EthPriorityOptionName) => {
+      setPriority(priority)
+      setReady() // clear error from previous submit attempt
+    },
+    [setPriority, setReady]
   )
 
   return (
@@ -172,37 +219,44 @@ export const EthSignTransactionRequest = () => {
         </div>
         {isReadyToDisplay && (
           <Suspense fallback={null}>
-            {nativeToken && transaction && txDetails && estimatedFee ? (
+            {transaction && txDetails && network?.nativeToken ? (
               <div className="gasInfo mt-8">
                 <div>
-                  <div>Estimated Fee</div>
+                  <div>
+                    Estimated Fee{" "}
+                    <WithTooltip
+                      tooltip={
+                        <FeeTooltip
+                          account={account?.address}
+                          tokenId={network.nativeToken.id}
+                          estimatedFee={txDetails.estimatedFee.toString()}
+                          maxFee={txDetails.maxFee.toString()}
+                        />
+                      }
+                    >
+                      <InfoIcon className="inline align-text-top" />
+                    </WithTooltip>
+                  </div>
                   <div>{transaction?.type === 2 && "Priority"}</div>
                 </div>
                 <div>
                   <div>
-                    <Tokens
-                      amount={estimatedFee.tokens}
-                      decimals={nativeToken.decimals}
-                      symbol={nativeToken.symbol}
-                      noCountUp
+                    <TokensAndFiat
+                      tokenId={network.nativeToken.id}
+                      planck={txDetails.estimatedFee.toString()}
                     />
-                    {estimatedFee && nativeTokenRates ? (
-                      <>
-                        {" "}
-                        (~
-                        <Fiat amount={estimatedFee.fiat("usd")} noCountUp currency="usd" />)
-                      </>
-                    ) : null}
                   </div>
                   <div>
                     <EthFeeSelect
+                      tx={request}
+                      tokenId={network.nativeToken.id}
                       disabled={isPayloadLocked}
-                      transaction={transaction}
+                      gasSettingsByPriority={gasSettingsByPriority}
+                      setCustomSettings={setCustomSettings}
                       txDetails={txDetails}
                       priority={priority}
-                      onChange={setPriority}
-                      decimals={nativeToken?.decimals}
-                      symbol={nativeToken?.symbol}
+                      onChange={handleFeeChange}
+                      networkUsage={networkUsage}
                     />
                   </div>
                 </div>
@@ -231,7 +285,7 @@ export const EthSignTransactionRequest = () => {
                   Cancel
                 </SimpleButton>
                 <SimpleButton
-                  disabled={!transaction || processing || isLoading}
+                  disabled={!transaction || processing || isLoading || !isValid}
                   processing={processing}
                   primary
                   onClick={approve}