+++ packages/balances-default-modules/src/index.test.ts
@@ -1,0 +1,5 @@
+export {}
+
+describe("Module", () => {
+  it("Can be implemented", () => {})
+})