+++ packages/chaindata-provider/src/types/Chain.ts
@@ -28,6 +28,8 @@ export type Chain = {
 
   paraId: number | null // The paraId of this chain, if it is a parachain
   relay?: Chain // The parent relayChain of this parachain, if this chain is a parachain
+
+  balanceMetadata: Array<{ moduleType: string; metadata: any }>
 }
 export type CustomChain = Chain & {
   isCustom: true