+++ packages/chain-connector-evm/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @talismn/chain-connector-evm
 
+## 0.4.0
+
+### Minor Changes
+
+- cd6a684: rotate rpcs on error
+
+### Patch Changes
+
+- 4aa691d: feat: new balance modules
+- Updated dependencies [4aa691d]
+  - @talismn/chaindata-provider@0.2.1
+
 ## 0.3.0
 
 ### Minor Changes